Add option to hide the review text from a Spice.
We have a helper function called {{starsAndReviews}} which actually accepts 4 arguments (the 4th one being the one that controls whether the review text is show or not) but it's not all being used. By passing meta.options.hideReviewText as the 4th argument, we'd be able to control the review text from a Spice / Goodie / whatever uses this template.